Add files via upload
These files are the AMPL run file which does a 5-period horizon correctly, and an output10 file. the other two are internal files created by the program which may be diagnostic. ETQ.mod, .dat are the quarter model declarations, the transport stage .mod, .dat are the scheduling model declarations.

the .docx is a general discussion of the strategy.